Remove path dependency, remove build script
This removes the build script (which didn't seem to be working), and
changes the dependency on syntex to be a git dependency.

This is needed for testing in a project that uses serde_macros.  The
project needs to link against both quasi without syntex (for
serde_macros) and quasi with syntex, so can't use a paths override in
.cargo/config.